vscode-extension/server/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
uvicorn>=0.30.0
fastapi>=0.112.0
python-dotenv>=1.0.1
pillow>=10.0.0
mss>=9.0.0
python-dateutil>=2.8.0
websockets>=12.0
pydantic>=2.0.0
